requirements.txt
flask
tensorflow
pillow
numpy 
kagglehub